requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.2
tweepy==4.14.0
requests==2.31.0
anthropic==0.3.0
groq==0.11.0
openai==1.12.0
firecrawl==2.31.0
Pillow==11.0.0
opencv-python==4.10.0
